+++ ai_tasks/future_tasks.md
@@ -1,0 +1,5 @@
+
+ - micromegas sql documentation
+ - rust tracing mode using feature: sync or async
+ - sql jit view
+ - async spans jit view based on async events view
