+++ rust/analytics/src/lib.rs
@@ -7,6 +7,8 @@
 pub mod arrow_properties;
 /// Misc arrow utilities
 pub mod arrow_utils;
+/// Processing of async events from thread event blocks
+pub mod async_block_processing;
 /// In-memory async events in arrow format
 pub mod async_events_table;
 /// Transforms thread events into call trees
